Move operation construction logic into a separate file
This change was done in preparation for adding new field handler logic
otherwise the main document class would have become cluttered up again.
All of the namespacing is still unchanged.
